+++ spec/dossier_spec.rb
@@ -28,30 +28,4 @@ describe Dossier do
     Dossier.configuration.should_receive(:client)
     Dossier.client
   end
-
-  describe "report naming" do
-    let(:klass) { HelloMyFriendsReport }
-    let(:name)  { 'hello_my_friends' }
-
-    it "converts a report class to a report name" do
-      expect(Dossier.class_to_name(klass)).to eq(name)
-    end
-
-    it "converting a report name to a report class" do
-      expect(Dossier.name_to_class(name)).to eq(klass)
-    end
-
-    describe "with namespaces" do
-      let(:klass) { Cats::Are::SuperFunReport }
-      let(:name)  { 'cats/are/super_fun' }
-
-      it "converts a report class to a report name" do
-        expect(Dossier.class_to_name klass).to eq name
-      end
-
-      it "converts a report name to a report class" do
-        expect(Dossier.name_to_class name).to eq klass
-      end
-    end
-  end
 end